refactor(api): move service to the services folder
